Fill PRINCIPLES.md expanding on the mw core REST principles
We still need to pull some of our principles and decisions from
the prototyping, but generally we agree with most of the mw core
principles and can just link to the text.

Perhaps we should link to a specific version of the text?
It probably doesnt matter much for the proposal as it won't be
changing much right now.