add auth and resulting error responses
This adds a global security scheme "bearerAuth" (which is optional) and
error responses to resources which can be affected. It provides a way
for users who are already authenticated (e.g. through OAuth or other
ways to become the "Bearer" of a token) to present their credentials. It
does not provide a solution for "Callers using cookie-based
authentication" as seen in the action API or mentioned in [0].

Generally, editing is possible as anonymous, but some resources (e.g.
[semi-protected] "pages"/entities) require elevated privileges and
attempts to manipulate their resources can lead to
authentication/authorization errors.

In contrast, to my knowledge, there is no existing way to read-protect
only some entities - consequently possible authentication/authorization
errors were only added to manipulations of resources.

See
https://swagger.io/docs/specification/authentication/bearer-authentication/

[0]: https://www.mediawiki.org/wiki/API:REST_API/Reference#Create_page